requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlmodel>=0.0.16,<1.0.0
asyncpg>=0.29.0,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
stripe>=8.0.0,<10.0.0
pydantic-settings>=2.1.0,<3.0.0
httpx>=0.27.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
google-genai>=0.2.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
apscheduler>=3.10.0,<4.0.0
email-validator>=2.1.0
python-dotenv>=1.0.0
